pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wizlight"
version = "0.1.0"
description = "Advanced WiZ smart light controller with screen sync, clap detection, and Alexa integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Naman"}
]
dependencies = [
    "pywizlight>=0.5.14",
    "click>=8.0",
    "mss>=9.0",
    "Pillow>=10.0",
    "customtkinter>=5.0",
    "sounddevice>=0.4",
    "numpy>=1.24",
    "fauxmo>=0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
wizlight = "src.cli.commands:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
